requirements.txt
flask
requests
google-generativeai
python-dotenv
flask-cors